added working screen-switching logic for menu.
