+++ core/src/com/mygdx/game/menuScreen.java
@@ -1,8 +1,11 @@
-package com.mygdx.game.menu;
+package com.mygdx.game;
 
 import com.badlogic.gdx.Gdx;
+import com.badlogic.gdx.Input;
 import com.badlogic.gdx.Screen;
 import com.badlogic.gdx.graphics.GL20;
+import com.badlogic.gdx.graphics.g2d.SpriteBatch;
+import com.badlogic.gdx.graphics.glutils.ShapeRenderer;
 import com.badlogic.gdx.scenes.scene2d.Stage;
 import com.badlogic.gdx.scenes.scene2d.ui.Skin;
 import com.badlogic.gdx.scenes.scene2d.ui.TextButton;
@@ -11,47 +14,56 @@ import com.mygdx.game.Omega;
 
 public class menuScreen implements Screen {
 
+    Omega game;
+
+    SpriteBatch mainBatch;
     private Stage stage;
     private Skin menuSkin;
     private TextButton PVP;
 
     public menuScreen(Omega game) {
         super();
-
-        menuSkin = new Skin(Gdx.files.internal("Skin/skin/vhs-ui.json"));
-
-        stage = new Stage(new FitViewport(1280, 720));
-        Gdx.input.setInputProcessor(stage);
-
-        PVP = new TextButton("Player vs Player", menuSkin);
-        PVP.setPosition(220, 50);
-        PVP.setSize(200, 100);
-
-        stage.addActor(PVP);
+        this.game = game;
+//        menuSkin = new Skin(Gdx.files.internal("Skin/skin/vhs-ui.json"));
+//
+//        stage = new Stage(new FitViewport(1280, 720));
+//        Gdx.input.setInputProcessor(stage);
+//
+//        PVP = new TextButton("Player vs Player", menuSkin);
+//        PVP.setPosition(220, 50);
+//        PVP.setSize(200, 100);
+//
+//        stage.addActor(PVP);
 
     }
 
     @Override
     public void render(float delta) {
-        Gdx.gl.glClearColor(0.4f, 0.5f, 0.8f, 1f);
+        Gdx.gl.glClearColor(0.1f, 0.1f, 0.1f, 1f);
         Gdx.gl.glClear(GL20.GL_COLOR_BUFFER_BIT);
-        stage.act();
-        stage.draw();
+        game.sr.begin(ShapeRenderer.ShapeType.Filled);
+        game.sr.rect(40,40,40,40);
+        game.sr.end();
+        if (Gdx.input.isKeyPressed(Input.Keys.SPACE)){
+            System.out.println("test");
+            this.dispose();
+            game.setScreen(new MainGameScreen(game));
+        }
     }
 
     @Override
     public void dispose() {
-        stage.dispose();
+
     }
 
     @Override
     public void show() {
-        Gdx.input.setInputProcessor(stage);
+
     }
 
     @Override
     public void hide() {
-        Gdx.input.setInputProcessor(null);
+
     }
 
     @Override
